Add SweetTicketDesign to pom.xml
issue #30

+++ src/main/java/pe/puyu/pukahttp/domain/FailedPrintJobsStorage.java
@@ -11,5 +11,5 @@ public interface FailedPrintJobsStorage {
     void deleteAll();
     void delete(PrintJob printJob);
     void deleteBefore(LocalDateTime beforeTime);
-    int countAll();
+    long countAll();
 }
